Add scheme choice hint to users + refactor bolt scheme utils
